// src/services/models/enderecoModel.ts
export interface Endereco {
    id_Endereco?: number;
    SB_Municipio: string;
    SB_Logradouro: string;
    SB_Numero: number;
    SB_Complemento: string;
    SB_Bairro: string;
    SB_ZonaPressao: string;
    SB_Polo: string;
    SB_Referencia: string;
    SB_SetorAbastecimento: string;
    SB_Cruzamento: string;
  }